+++ lilo/src/main/java/com/amrdeveloper/lilo/object/LiloFloat.kt
@@ -7,14 +7,14 @@ import com.amrdeveloper.lilo.runtime.LiloInterpreter
 import com.amrdeveloper.lilo.type.liloFloatType
 import com.amrdeveloper.lilo.type.liloMethodType
 
-data class LiloFloat(val value: Float) : LiloObject(liloFloatType) {
+data class LiloFloat(val value: Double) : LiloObject(liloFloatType) {
 
     init {
         setAttr(name = LiloMagicMethod.BOOL, value = FloatBool)
     }
 
     override fun toString(): String {
-        if (value == Float.POSITIVE_INFINITY) return "inf"
+        if (value == Double.POSITIVE_INFINITY) return "inf"
         if (value.isNaN()) return "nan"
         return value.toString()
     }
@@ -26,6 +26,6 @@ private object FloatBool : LiloObject(liloMethodType), LiloCallable {
         args: List<LiloObject>
     ): LiloResult<LiloObject> {
         val self = args[0] as LiloFloat
-        return LiloResult.Success(data = LiloBool(value = self.value != 0.0f))
+        return LiloResult.Success(data = LiloBool(value = self.value != 0.0))
     }
 }
